+++ app/routes.ts
@@ -1,8 +1,5 @@
 import { type RouteConfig, layout, route } from "@react-router/dev/routes";
 
 export default [
-  layout("./layout.tsx", [
-    route("/", "routes/home.tsx"),
-    route("/boards", "routes/boards.tsx"),
-  ]),
+  layout("./layout.tsx", [route("/", "routes/home.tsx")]),
 ] satisfies RouteConfig;
